requirements.txt
azure-functions
fastapi
uvicorn
selenium
beautifulsoup4
webdriver_manager
lxml